Work-in-Progress Update
Adding the patch system, struggling to find the download links for them so it requires some time, as well as the patch system not being fully implemented yet, so I'm committing out of a hurry. Outside of that, I divided the files for more order so that the entire core is not stored entirely within one file. Keep on using the Port-Blocking option as the patch option is still in WIP