Comment clean-up
I could not push an update today due to unforeseen temporal issues. Regular updates will begin tomorrow. I wanted to remove some of the more useless and inappropriate comments for the sake of professionalism. Regardless, I can't deny that this project is coming to an end and that I'll have to work on a new one soon enough. Don't worry, I have some vague ideas on what to make to keep the schedule up. I'll probably work on a new project every week. If I don't finish the project by the end of the week, I'll move to a new one and try to maintain the older project, even though all of my attention will be aimed toward the newer project.